+++ src/api/user.rs
@@ -1,0 +1,44 @@
+/*     
+ * ripalt
+ * Copyright (C) 2018 Daniel Müller
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+use super::*;
+use handlers::user::UserStats;
+use models::user::UserStatsMsg;
+
+pub fn stats(req: HttpRequest<State>) -> FutureResponse<HttpResponse> {
+    if let Some(identity) = req.identity() {
+        let user_id = match Uuid::parse_str(identity).map_err(|e| ErrorInternalServerError(e)) {
+            Ok(user_id) => user_id,
+            Err(e) => return Box::new(FutErr(e)),
+        };
+
+        req.state().db().send(UserStats(user_id))
+            .from_err()
+            .and_then(|result: Result<UserStatsMsg>| {
+                match result {
+                    Ok(stats) => {
+                        Ok(HttpResponse::Ok().json(stats))
+                    },
+                    Err(_) => Ok(HttpResponse::InternalServerError().into()),
+                }
+            })
+            .responder()
+    } else {
+        Box::new(FutErr(ErrorUnauthorized("unauthorized")))
+    }
+}