+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Cleanup thread to remove orphaned peers.
 - Identity Provider for the API, which uses either the Session ID or a JWT.
     - New Setting: `jwt_secret`, the secret key for the JWTs.
+- API Endpoint to get the own stats (/api/v1/user/stats)
 
 ## [0.1.0]